update (добавил возможность вывода загрузки)

+++ Console_manipulation.cpp
@@ -160,7 +160,7 @@ void Console_manipulation::load_saved_cursor_pos()
 	}
 }
 
-Console_show_loading::Console_show_loading(int start_value, int finish_value)
+Console_show_loading::Console_show_loading(unsigned long long start_value, unsigned long long finish_value)
 {
 	this->start_value = start_value;
 	this->finish_value = finish_value;
@@ -173,21 +173,21 @@ Console_show_loading::Console_show_loading(int start_value, int finish_value)
 	return;
 }
 
-Console_show_loading::Console_show_loading(int start_value, int finish_value, int max_quantity_divisions) : Console_show_loading(start_value, finish_value)
+Console_show_loading::Console_show_loading(unsigned long long start_value, unsigned long long finish_value, unsigned long long max_quantity_divisions) : Console_show_loading(start_value, finish_value)
 {
 	this->max_quantity_divisions = max_quantity_divisions;
 }
 
-Console_show_loading::Console_show_loading(int start_value, int finish_value, char filler, int max_quantity_divisions) : Console_show_loading(start_value, finish_value, max_quantity_divisions)
+Console_show_loading::Console_show_loading(unsigned long long start_value, unsigned long long finish_value, char filler, unsigned long long max_quantity_divisions) : Console_show_loading(start_value, finish_value, max_quantity_divisions)
 {
 	this->filler = filler;
 }
-Console_show_loading::Console_show_loading(int start_value, int finish_value, char filler, int max_quantity_divisions, Text_color color) : Console_show_loading(start_value, finish_value, filler, max_quantity_divisions)
+Console_show_loading::Console_show_loading(unsigned long long start_value, unsigned long long finish_value, char filler, unsigned long long max_quantity_divisions, Text_color color) : Console_show_loading(start_value, finish_value, filler, max_quantity_divisions)
 {
 	this->color = color;
 }
 
-void Console_show_loading::show_loading(const int now_value) const
+void Console_show_loading::show_loading(const unsigned long long now_value) const
 {
 	std::cout << "{";
