+++ Console_manipulation.h
@@ -99,17 +99,17 @@ class Console_show_loading
 {
 public:
 
-	Console_show_loading(int start_value, int finish_value);
-	Console_show_loading(int start_value, int finish_value, int max_quantity_divisions);
-	Console_show_loading(int start_value, int finish_value, char filler, int max_quantity_divisions);
-	Console_show_loading(int start_value, int finish_value, char filler, int max_quantity_divisions, Text_color color);
+	Console_show_loading(unsigned long long start_value, unsigned long long finish_value);
+	Console_show_loading(unsigned long long start_value, unsigned long long finish_value, unsigned long long max_quantity_divisions);
+	Console_show_loading(unsigned long long start_value, unsigned long long finish_value, char filler, unsigned long long max_quantity_divisions);
+	Console_show_loading(unsigned long long start_value, unsigned long long finish_value, char filler, unsigned long long max_quantity_divisions, Text_color color);
 
-	void show_loading(const int now_value) const;
+	void show_loading(const unsigned long long now_value) const;
 
 private:
 
-	int start_value, finish_value;
-	int max_quantity_divisions;
+	unsigned long long start_value, finish_value;
+	unsigned long long max_quantity_divisions;
 	Text_color color;
 	char filler;
 };
